backend/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
config==0.5.1
dlib==19.24.6
Flask==2.2.5
Flask-JWT-Extended==4.6.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
psycopg==3.2.2
pycparser==2.22
PyJWT==2.10.0
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
